FraudDetectionAgent/src/FraudDetectionAgent/agent.py

from autogen import AssistantAgent
from agoric_sdk import ContractAPI

class FraudDetectionAgent:
    def __init__(self):
        self.assistant = AssistantAgent("fraud_detection_assistant")
        self.contract_api = ContractAPI("https://api.agoric.net")

    def analyze_transactions(self, address):
        try:
            # Analyze transactions for fraud detection
            transactions = self.contract_api.get_transactions(address)
            fraud_analysis = self.assistant.handle_input(f"Analyze the following transactions for fraud: {transactions}")
            return fraud_analysis
        except Exception as e:
            return f"Failed to analyze transactions: {str(e)}"

    def send_alert(self, message):
        try:
            # Send an alert
            alert = self.assistant.handle_input(message)
            return alert
        except Exception as e:
            return f"Failed to send alert: {str(e)}"

    def generate_fraud_report(self, address):
        try:
            # Generate a fraud report for the given address
            transactions = self.contract_api.get_transactions(address)
            report = self.assistant.handle_input(f"Generate a fraud report for the following transactions: {transactions}")
            return report
        except Exception as e:
            return f"Failed to generate fraud report: {str(e)}"

if __name__ == "__main__":
    agent = FraudDetectionAgent()
    address = "example_address"
    fraud_analysis_result = agent.analyze_transactions(address)
    print(fraud_analysis_result)
    alert_result = agent.send_alert("Potential fraud detected!")
    print(alert_result)
    fraud_report_result = agent.generate_fraud_report(address)
    print(fraud_report_result)
